Rifactoring error cases

+++ system/application/helpers/actions_helper.php
@@ -12,7 +12,7 @@ function get_action_strings($actions, $tuple) {
 	foreach ($actions as $action) {
 		switch ($action) {
 			case 'profile':
-				$str .= '<li><a href="/profile/user/'.$tuple['account_id'].'" class="ajaxlink">See Profile</a></li>';
+				$str .= '<li><a href="/profile/user/'.$tuple['account_id'].'" class="ajax">See Profile</a></li>';
 				break;
 			
 			case 'send-email':
@@ -33,19 +33,19 @@ function get_action_strings($actions, $tuple) {
 				break;	
 
 			case 'delete-conn':
-				$str .= '<li><a href="/connections/destroy/'.$tuple['account_id'].'">Delete Connection</a></li>';
+				$str .= '<li><a href="/connections/destroy/'.$tuple['account_id'].'" class="ajax-confirm">Delete Connection</a></li>';
 				break;
 			case 'accept-conn-req':
-				$str .= '<li><a href="/connections/accept/'.$tuple['account_id'].'">Accept Connection</a></li>';
+				$str .= '<li><a href="/connections/accept/'.$tuple['account_id'].'" class="ajax-confirm">Accept Connection</a></li>';
 				break;
 			case 'reject-conn-req':
-				$str .= '<li><a href="/connections/reject/'.$tuple['account_id'].'">Reject Connection</a></li>';
+				$str .= '<li><a href="/connections/reject/'.$tuple['account_id'].'" class="ajax-confirm">Reject Connection</a></li>';
 				break;
 			case 'cancel-conn-req':
-				$str .= '<li><a href="/connections/cancel/'.$tuple['account_id'].'">Cancel Connection Request</a></li>';
+				$str .= '<li><a href="/connections/cancel/'.$tuple['account_id'].'" class="ajax-confirm">Cancel Connection Request</a></li>';
 				break;
 			case 'request-conn':
-				$str .= '<li><a href="/connections/request/'.$tuple['account_id'].'">Request Connection</a></li>';
+				$str .= '<li><a href="/connections/request/'.$tuple['account_id'].'" class=\"ajax-confirm\">Request Connection</a></li>';
 				break;
 			
 			case 'issue-bill':
